% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByGroup.R
\name{plotByGroup}
\alias{plotByGroup}
\title{Visualize one categorical column and one numeric column}
\usage{
plotByGroup(x, x.factor = NULL, y.numeric = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{x.factor}{A column in \code{sample_data} to compare that is a factor}

\item{y.numeric}{A column in \code{sample_data} to compare that is a numeric}
}
\value{
a \code{ggplot} object.
}
\description{
Visualize one categorical column and one numeric column
}
\details{
A simple plotting utility for visualization of one categorical
         and numerical variable. Provided a categorical variable and numerical
         variable present in \code{sample_data} of a
         \code{\link{phyloseq-class}} object, a comparative visualization
         can be done.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(dplyr)
library(microbiome)
library(ggplot2)
ps <- FuentesIliGutData
# calculate Shannon diversity using microbiome::diversity and add it to
# the sample_data in phyloseq using biomeUtils::mutateSampleData()
ps <- mutateSampleData(ps,
                       Shannon = microbiome::diversity(ps, "shannon")[,1])
plotByGroup(ps,
            x.factor="ILI",
            y.numeric = "Shannon") +
         geom_point(aes(color=ILI),
                    position = position_jitter(width = 0.1)) +
         scale_color_manual(values=c("steelblue", "brown3", "grey70"))

}
\references{
Shetty SA (2021). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeViz}
}
\author{
Sudarshan A. Shetty
}
