% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCoA.R
\name{plotPCoA}
\alias{plotPCoA}
\title{PCoA plot}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{group_var}{A column in \code{sample_data} to compare. This is also the
variable used to colour.}

\item{ord_method}{Ordination method, currently tested PCoA.}

\item{dist_method}{Distance method, currently tested Bray-Curtis.}

\item{seed}{Random seed number \code{set.seed}.}

\item{cor_method}{Correlation method. Default is \code{Spearman}.}

\item{padj_cutoff}{Cut-off for multiple testing. Default is 0.05.}

\item{padj_method}{Method for multiple testing. Default is fdr.}

\item{arrows}{Logical. If arrows for taxa with correlation to axis are to be
plotted. Default is TRUE.}

\item{label_col}{Color of labels. Default is grey30.}

\item{plot_centroids}{Logical. To plot centroids or not. Default is TRUE.}

\item{add_side_box}{Logical. To plot side boxplots or not. Default is TRUE.}

\item{axis_plot}{Which axis to plot. Default is first two.}

\item{point_shape}{Shape of the points. Default is 21.}

\item{point_alpha}{Opacity of points. Default is 0.5}

\item{verbose}{Logical. Messages to print. Default is TRUE.}

\item{...}{Additional arguments to pass to vegan's adonis function.}
}
\value{
a \code{ggplot2} object.
}
\description{
PCoA plot
}
\details{
A Principal Coordinates Analysis for \code{\link{phyloseq-class}} object.
         To visualize similarities/dissimilarities between samples in 2D ordination.
         This function extends the \code{phyloseq} ordination plots to include
         taxa that correlate with choosen axis and plots them along with a
         side boxplot for comparing inter-sample variation within groups.
}
\examples{
library(biomeUtils)
library(dplyr)
library(ggside)
ps <- FuentesIliGutData \%>\%
  microbiome::transform("compositional") \%>\%
  mutateTaxaTable(FeatureID = taxa_names(FuentesIliGutData))
plotPCoA(x =ps,
         group_var = "ILI",
         ord_method = "PCoA",
         dist_method = "bray",
         seed = 1253,
         cor_method = "spearman",
         verbose = TRUE,
         padj_cutoff = 0.05,
         padj_method = "fdr",
         arrows = TRUE,
         label_col = "grey30",
         plot_centroids = TRUE,
         add_side_box = TRUE,
         axis_plot = c(1:2),
         point_shape = 21,  # point_shape
         point_alpha = 0.5) +
  scale_color_manual(values = c("#3d6721", "#a86826", "#006c89")) +
  scale_fill_manual(values = c("#3d6721", "#a86826", "#006c89"))

}
\references{
Shetty SA (2021). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeViz}
}
\author{
Sudarshan A. Shetty
}
