% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterViz.R
\name{plotScatterViz}
\alias{plotScatterViz}
\title{Scater Plot}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{x_numeric}{A column in \code{sample_data} to compare that is a numeric}

\item{y_numeric}{A column in \code{sample_data} to compare that is a numeric}

\item{color_var}{A column in \code{sample_data} to color points}

\item{point_size}{\code{geom_point}. Default is 3}

\item{point_opacity}{\code{ggplot2} alpha value. Default is 0.5}

\item{point_shape}{\code{ggplot2} point shape value. Default is 21}

\item{side_panel}{Logical. If side density to plot. Default is TRUE}

\item{side_panel_scale}{\code{ggside} ggside.panel.scale requires a scalar numeric.
default is 0.3}
}
\value{
a \code{ggplot} object.
}
\description{
Scater Plot
}
\details{
A simple plotting utility for visualization of two numerical
         variables. Provided two numerical
         variable present in \code{sample_data} of a
         \code{\link{phyloseq-class}} object, a scatter plot is returned.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(dplyr)
library(microbiome)
library(ggplot2)
ps <- FuentesIliGutData
plotScatterViz(ps, x_numeric = "age", y_numeric = "BMI",
               color_var = "ILI") +
               scale_fill_manual(values=c("steelblue", "brown3", "grey70"))+
               scale_fill_manual(values=c("steelblue", "brown3", "grey70"))
}
\references{
Shetty SA (2021). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeViz}
}
\author{
Sudarshan A. Shetty
}
