% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTaxaRaincloud.R
\name{plotTaxaRaincloud}
\alias{plotTaxaRaincloud}
\title{Creates a Raincloud plot}
\usage{
plotTaxaRaincloud(
  x,
  taxa = NULL,
  group_samples_by = NULL,
  opacity = 0.25,
  shape_point = 21
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{taxa}{Feature to plot.}

\item{group_samples_by}{A column in \code{sample_data} to compare abundances.}

\item{opacity}{Alpha value.}

\item{shape_point}{Point shape.}
}
\value{
a \code{ggplot2} object.
}
\description{
Creates a Raincloud plot
}
\details{
A Rain Cloud plotting wrapper for \code{\link{phyloseq-class}} object.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(microbiome)
library(dplyr)

ps <- FuentesIliGutData \%>\%
       microbiome::aggregate_taxa("Genus") \%>\%
       microbiome::transform("compositional")

# Select taxa and group to compare

plotTaxaRaincloud(ps,
                  taxa ="Bacteroides",
                  group_samples_by = "ILI",
                  opacity = 0.25,
                  shape_point = 21)

}
\references{
Shetty SA (2021). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeViz}

Scherer Cédric (2021). A Step-by-Step tutorial as supplement to my talk “ggplot
Wizardry: My Favorite Tricks and Secrets for Beautiful Plot in R” at OutlierConf 2021.
\url{https://z3tt.github.io/OutlierConf2021/}
}
\author{
Sudarshan A. Shetty
}
