% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTidyHeatmap.R
\name{plotTidyHeatmap}
\alias{plotTidyHeatmap}
\title{Creates a Heatmap using `TidyHeatmap` and `ComplexHeatmap`}
\usage{
plotTidyHeatmap(
  x,
  select_taxa = NULL,
  group_samples_by = NULL,
  add_taxa_label = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{select_taxa}{Features to plot.}

\item{group_samples_by}{A column in \code{sample_data} to group tiles.}

\item{add_taxa_label}{Logical. Default is TRUE.}

\item{...}{Arguments forwarded to tidyHeatmap::heatmap and ComplexHeatmap::Heatmap.}
}
\value{
a \code{heatmap} object.
}
\description{
Creates a Heatmap using `TidyHeatmap` and `ComplexHeatmap`
}
\details{
A TidyHeatmap plotting wrapper for \code{\link{phyloseq-class}} object.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(microbiome)
library(dplyr)

ps <- FuentesIliGutData \%>\%
         microbiome::transform("compositional")
select_taxa <- findTopTaxa(ps, top= 10, method="mean")

plotTidyHeatmap(ps, select_taxa = select_taxa,
                group_samples_by = "ILI",
                add_taxa_label = FALSE,
                cluster_rows = FALSE,
                .scale = "none",
                transform = NULL,
                palette_value = c("red", "white", "blue"))

}
\references{
Shetty SA (2021). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeViz}
}
\author{
Sudarshan A. Shetty
}
