% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopAbundant.R
\name{plotTopAbundant}
\alias{plotTopAbundant}
\title{Plot Top Abundant Taxa}
\usage{
plotTopAbundant(x, taxa_level = "Family", top = 10L, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{taxa_level}{Taxonomic level at which features are selected. Abundances
are summed up for each of the taxa at this level and then top
are selected.}

\item{top}{Numeric value, how many top \code{taxa_level} to return. Default
top = 100}

\item{...}{Options to pass ggplot2::geom_jitter()}
}
\value{
A ggplot2 object.
}
\description{
Plot Top Abundant Taxa
}
\details{
Plot the features with the highest abundance in all samples,
         at a user specific taxonomic level.
}
\examples{

library(biomeUtils)
library(biomeViz)
p <- plotTopAbundant(SprockettTHData,
                     taxa_level = "Family",
                     top=10L,
                     alpha = 0.25,
                     shape = 124,
                     color="black",
                     fill = "black",
                     size=3,
                     width = 0.02)
print(p)

}
\references{
Shetty SA (2020). Data visualization for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
