% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_biomViz}
\alias{theme_biomViz}
\title{Standard ggplot2 theme for RIVM Toolbox}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base family}
}
\description{
Standard ggplot2 theme for RIVM Toolbox
}
\details{
A standard theme for ggplot2 to produce similar aesthetics for
         RIVM Toolbox.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(microbiome)
library(ggplot2)
library(dplyr)

ps <- FuentesIliGutData \%>\%
  microbiome::aggregate_taxa("Genus") \%>\%
  microbiome::transform("compositional")

plotTaxaRaincloud(ps,
                  taxa ="Bacteroides",
                  group_samples_by = "ILI",
                  opacity = 0.25,
                  shape_point = 21) + # combine with ggplot2 for improvements
  labs(y = "Relative abundance",
       x = "",
       subtitle = "Raincloud plot") +
  theme_biomViz()+
  scale_fill_manual(values = c("#3d6721", "#a86826", "#006c89"), guide = "none") +
  scale_color_manual(values = c("#3d6721", "#a86826", "#006c89"), guide = "none") +
  theme(plot.subtitle = element_text(margin = margin(t = 5, b = 10)),
        plot.margin = margin(10, 25, 10, 25))

}
