% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Themes}
\alias{Themes}
\alias{theme_biomViz}
\alias{theme_biomViz_minimal}
\alias{theme_biomViz_bw}
\alias{scale_fill_biomeViz_summer}
\alias{scale_colour_biomeViz_summer}
\title{Standard ggplot2 theme for RIVM Toolbox}
\usage{
theme_biomViz(base_size = 11, base_family = "")

theme_biomViz_minimal(base_size = 11, base_family = "")

theme_biomViz_bw(base_size = 11, base_family = "")

scale_fill_biomeViz_summer(...)

scale_colour_biomeViz_summer(...)
}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base family}
}
\description{
Standard ggplot2 theme for RIVM Toolbox
}
\details{
A standard theme for ggplot2 to produce similar aesthetics for
         RIVM Toolbox.
}
\examples{
library(biomeUtils)
library(biomeViz)
library(microbiome)
library(ggplot2)
library(dplyr)

ps <- FuentesIliGutData \%>\%
  microbiome::aggregate_taxa("Genus") \%>\%
  microbiome::transform("compositional")

plotTaxaRaincloud(ps,
                  taxa ="Bacteroides",
                  group_samples_by = "ILI",
                  opacity = 0.25,
                  shape_point = 21) + # combine with ggplot2 for improvements
  labs(y = "Relative abundance",
       x = "",
       subtitle = "Raincloud plot") +
  theme_biomViz_minimal()+
  scale_fill_biomeViz_summer() +
  scale_colour_biomeViz_summer()

}
